requirements.txt
Django~=4.2.1
requests~=2.30.0
djangorestframework~=3.14.0